# src/main/resources/openapi.yaml
openapi: "3.0.3"
info:
  title: "TV_Program_Plan_Frequency API"
  description: "TV_Program_Plan_Frequency API"
  version: "1.0.0"
servers:
  - url: "https://TV_Program_Plan_Frequency"
paths:
  /api/mostFrequantShows/{date}:
    get:
      summary: "GET api/mostFrequantShows/{date}"
      operationId: "getMostFrequentShows"
      parameters:
        - name: "date"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/mostFrequantShows:
    get:
      summary: "GET api/mostFrequantShows"
      operationId: "getMostFrequentShowsWithoutDate"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/frequencyOrderedShows/{date}:
    get:
      summary: "GET api/frequencyOrderedShows/{date}"
      operationId: "getFrequencyOrderedShows"
      parameters:
        - name: "date"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/frequencyOrderedShows:
    get:
      summary: "GET api/frequencyOrderedShows"
      operationId: "getFrequencyOrderedShows"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/mostFrequantShowsOrderedShowsByType/{date}/{type}:
    get:
      summary: "GET api/mostFrequantShowsOrderedShowsByType/{date}/{type}"
      operationId: "getMostFrequentShowsByShowType"
      parameters:
        - name: "date"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "type"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "TVSHOW"
              - "SERIES"
              - "MOVIE"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/mostFrequantShowsOrderedShowsByType/{type}:
    get:
      summary: "GET api/mostFrequantShowsOrderedShowsByType/{type}"
      operationId: "getMostFrequentShowsByShowType"
      parameters:
        - name: "type"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "TVSHOW"
              - "SERIES"
              - "MOVIE"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/frequencyOrderedShowsByType/{date}/{type}:
    get:
      summary: "GET api/frequencyOrderedShowsByType/{date}/{type}"
      operationId: "getFrequencyOrderedShowsByType"
      parameters:
        - name: "date"
          in: "path"
          required: true
          schema:
            type: "string"
        - name: "type"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "TVSHOW"
              - "SERIES"
              - "MOVIE"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/frequencyOrderedShowsByType/{type}:
    get:
      summary: "GET api/frequencyOrderedShowsByType/{type}"
      operationId: "getFrequencyOrderedShowsByTypeWithoutDate"
      parameters:
        - name: "type"
          in: "path"
          required: true
          schema:
            type: "string"
            enum:
              - "TVSHOW"
              - "SERIES"
              - "MOVIE"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/FrequencyResponse"
  /api/SeriesByStatus/{date}:
    get:
      summary: "GET api/SeriesByStatus/{date}"
      operationId: "getSeriesByStatus"
      parameters:
        - name: "date"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/StatusResponse"
  /api/SeriesByStatus:
    get:
      summary: "GET api/SeriesByStatus"
      operationId: "getSeriesByStatus"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/StatusResponse"
components:
  schemas:
    FrequencyResponse:
      type: "object"
      properties:
        title:
          type: "string"
        frequency:
          type: "integer"
          format: "int32"
    StatusResponse:
      type: "object"
      properties:
        title:
          type: "string"
        status:
          type: "string"
          enum:
            - "RUNNING"
            - "ENDED"